Add org readme repo
Adding a .github repo

The intention is to be able to have a README on org level, i.e. so
that a readme can be shown if you go to https://github.com/eclipse-kuksa.
That could be used to provide a guide to the repos.

Intention is to have something similar to eclipse-velocitas.
If you go to https://github.com/eclipse-velocitas
you will see a readme.
That readme comes from
https://github.com/eclipse-velocitas/.github/blob/main/profile/README.md