Add repos for kuksa.val.services migration
Intended for migration of https://github.com/eclipse/kuksa.val.services/

hvac_service and seat_service to be moved to kuksa-incubation.
They are considered as examples and should not be used for production purposes.
For kuksa-incubation we intend to not have dependabot alerts or PRs,
to minimize maintenance needs,

mock_service intended to be migrated to kuksa-mock-provider